requirements.txt
streamlit
pandas
plotly
biopython
stmol
py3Dmol
ipython
ipython_genutils